+++ src/main/java/fr/mkadia/mkadiaapi/repositories/MediaRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MediaRepository extends JpaRepository<Media , Integer> {
+public interface MediaRepository extends JpaRepository<Media , Long> {
     List<Media> findAllByProduct(Product product);
 }
