+++ src/main/java/fr/mkadia/mkadiaapi/repositories/ReviewRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
-    Page<Review> findByProductIdOrderByCreatedAtDesc(Integer productId, Pageable pageable);
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+    Page<Review> findByProductIdOrderByCreatedAtDesc(Long productId, Pageable pageable);
 }
