+++ src/main/resources/db/migration/V1_18__create_favorites_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE favorites
 (
-    favorite_id         INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-    user_id    INT    NOT NULL,
-    product_id INT    NOT NULL,
+    favorite_id         BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    user_id    BIGINT    NOT NULL,
+    product_id BIGINT    NOT NULL,
 
     created_at   TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at   TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
